backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-unfold>=0.74.1",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg[binary]>=3.3.2",
    "whitenoise>=6.11.0",
]
